+++ .prettierrc
@@ -1,0 +1,9 @@
+{
+    "semi": true,
+    "tabWidth": 2,
+    "printWidth": 80,
+    "singleQuote": false,
+    "trailingComma": "all",
+    "useTabs": false
+  }
+
