+++ src/components/Explore/Explore.jsx
@@ -8,8 +8,8 @@ function Explore() {
   return (
     <div className="explore">
       <h1 style={{ textAlign: 'center' }}>Explore our Mentors</h1>
-     <Search />
-     <ExploreGrid />
+      <Search />
+      <ExploreGrid />
     </div>
   );
 }
