--- src/index.js
@@ -1,27 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import { BrowserRouter } from "react-router-dom";
-
-import reportWebVitals from './reportWebVitals';
-import {createStore} from '@reduxjs/toolkit';
-import {Provider} from 'react-redux';
-import rootReducer from './store/reducers/rootReducer';
-
-const store = createStore(rootReducer);
-
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>
-</React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
