+++ src/index.jsx
@@ -1,0 +1,25 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import { BrowserRouter } from "react-router-dom";
+import { createStore } from "@reduxjs/toolkit";
+import App from "./App";
+
+import reportWebVitals from "./reportWebVitals";
+import rootReducer from "./store/reducers/rootReducer";
+
+// const store = createStore(rootReducer);
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>,
+);
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
