+++ src/modules/FindAMentor/FindAMentor.jsx
@@ -1,0 +1,14 @@
+import ExploreGrid from '../../components/ExploreGrid/ExploreGrid';
+import Header from '../../components/Header/Header';
+import Search from '../../components/Search/Search';
+
+export default function FindAMentor() {
+  return (
+    <>
+      <Header />
+      <div></div>
+      <Search />
+      <ExploreGrid />
+    </>
+  );
+}
