--- src/modules/Login/Login.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Login() {
-  return (
-    <div>Login</div>
-  )
-}
-
-export default Login
