--- src/modules/Register/Register.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Register() {
-  return (
-    <div>Register</div>
-  )
-}
-
-export default Register
